Choose article by vehicle selections

+++ src/main/java/com/automaterijal/application/domain/entity/tecdoc/TecDocAtributi.java
@@ -1,29 +1,29 @@
 package com.automaterijal.application.domain.entity.tecdoc;
 
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
-
-import jakarta.persistence.*;
 
 @Entity
 @Data
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class TecDocAtributi {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column
-    Integer id;
-    Integer tecDocArticleId;
-    Long robaId;
-    String ppid;
-    Integer tecDocPpid;
-    String katbr;
-    String attrValue;
-    String attrUnit;
-    String attrShortName;
-    String attrType;
-    String dokumentId;
-    byte[] dokument;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column
+  Integer id;
+
+  Long tecDocArticleId;
+  Long robaId;
+  String ppid;
+  Long tecDocPpid;
+  String katbr;
+  String attrValue;
+  String attrUnit;
+  String attrShortName;
+  String attrType;
+  String dokumentId;
+  byte[] dokument;
 }
